+++ packages/form/src/demos/components-other.tsx
@@ -12,6 +12,7 @@ import ProForm, {
   ProFormDigit,
   ProFormDateTimePicker,
   ProFormSlider,
+  ProFormDateTimeRangePicker,
   ProFormDateRangePicker,
   ProFormUploadButton,
   ProFormUploadDragger,
@@ -109,7 +110,6 @@ const Demo = () => (
         action="/upload.do"
         extra="longgggggggggggggggggggggggggggggggggg"
       />
-
       <ProForm.Group title="日期相关分组">
         <ProFormDatePicker name="date" label="日期" />
         <ProFormDatePicker.Week name="dateWeek" label="周" />
@@ -118,6 +118,7 @@ const Demo = () => (
         <ProFormDatePicker.Year name="dateYear" label="年" />
         <ProFormDateTimePicker name="dateTime" label="日期时间" />
         <ProFormDateRangePicker name="dateRange" label="日期区间" />
+        <ProFormDateTimeRangePicker name="dateTimeRange" label="日期时间区间" />
       </ProForm.Group>
       <ProFormUploadDragger label="Dragger" name="dragger" />
     </ProForm>