+++ packages/utils/src/index.tsx
@@ -7,9 +7,10 @@ import BatchOperationBar from './components/BatchOperationBar';
 import isBrowser from './isBrowser';
 import isImg from './isImg';
 import isUrl from './isUrl';
+import isNil from './isNil';
 import pickProProps from './pickProProps';
-import pickUndefined from './pickUndefined';
-import pickUndefinedAndArray from './pickUndefinedAndArray';
+import omitUndefined from './omitUndefined';
+import omitUndefinedAndEmptyArr from './omitUndefinedAndEmptyArr';
 import pickProFormItemProps from './pickProFormItemProps';
 
 /**
@@ -52,8 +53,9 @@ export {
   parseValueToMoment,
   useDocumentTitle,
   isImg,
-  pickUndefined,
-  pickUndefinedAndArray,
+  isNil,
+  omitUndefined,
+  omitUndefinedAndEmptyArr,
   pickProFormItemProps,
   isUrl,
   isBrowser,