+++ packages/utils/src/isNil/index.ts
@@ -1,0 +1,3 @@
+const isNil = (value: any) => value === null || value === undefined;
+
+export default isNil;