+++ tests/utils/index.test.ts
@@ -1,4 +1,4 @@
-import { conversionSubmitValue, parseValueToMoment } from '@ant-design/pro-utils';
+import { conversionSubmitValue, parseValueToMoment, isNil } from '@ant-design/pro-utils';
 import moment, { Moment } from 'moment';
 
 describe('utils', () => {
@@ -16,7 +16,7 @@ describe('utils', () => {
       },
       'string',
       {
-        value: 'dataTime',
+        dataTime: 'dataTime',
         time: 'time',
         name: 'text',
         dateRange: 'dateRange',
@@ -46,7 +46,7 @@ describe('utils', () => {
       },
       'number',
       {
-        value: 'dataTime',
+        dateTime: 'dataTime',
         time: 'time',
         name: 'text',
         dateRange: 'dateRange',
@@ -76,7 +76,7 @@ describe('utils', () => {
       },
       false,
       {
-        value: 'dataTime',
+        dateTime: 'dataTime',
         time: 'time',
         name: 'text',
         dateRange: 'dateRange',
@@ -104,5 +104,14 @@ describe('utils', () => {
     expect((html as Moment[]).map((item) => item.valueOf()).join(',')).toBe(
       '1573862400000,1573862400000',
     );
+  });
+
+  it('📅 isNil', async () => {
+    expect(isNil(null)).toBe(true);
+    expect(isNil(undefined)).toBe(true);
+    expect(isNil(0)).toBe(false);
+    expect(isNil('')).toBe(false);
+    expect(isNil({})).toBe(false);
+    expect(isNil(true)).toBe(false);
   });
 });