+++ packages/layout/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to this project will be documented in this file. See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.4.17](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.4.16...@ant-design/pro-layout@6.4.17) (2020-09-14)
+
+**Note:** Version bump only for package @ant-design/pro-layout
+
+## [6.4.16](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.4.15...@ant-design/pro-layout@6.4.16) (2020-09-10)
+
+### Bug Fixes
+
+- **layout:** support using GlobalHeader prefixCls ([#411](https://github.com/ant-design/pro-components/issues/411)) ([d820fa3](https://github.com/ant-design/pro-components/commit/d820fa3aa75385b2262fe3e70101ab3664eea44f))
+
 ## [6.4.15](https://github.com/ant-design/pro-components/compare/@ant-design/pro-layout@6.4.14...@ant-design/pro-layout@6.4.15) (2020-09-07)
 
 ### Bug Fixes